backend/requirements.txt
Django==5.2
django-filter==24.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.12
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0